Move event preparation RH to same file as creation
